Pequeños cambios

+++ src/ejerciciospolotic_tp2/Ejercicio04.java
@@ -10,9 +10,9 @@ public class Ejercicio04 {
     public static void main(String[] args) {
         System.out.println("Ingresar el valor del cual se quiere calcular su factorial");
         Scanner n = new Scanner(System.in);
-        int valor = n.nextInt();
-        int base = valor;
-        int factorial=valor;
+        double valor = n.nextInt();
+        double base = valor;
+        double factorial=valor;
         while (valor>1){
             factorial=factorial*(valor-1);
             valor--;
